Extract some parts of code from AbstractShopManager
